Add third column to master_evaluation_modify table
It was already added in some tests, but now make it globally
applicable to the test file.

+++ src/test/regress/sql/master_evaluation_modify.sql
@@ -38,7 +38,7 @@ CREATE TYPE user_data AS (name text, age int);
 SET citus.replication_model TO streaming;
 SET citus.shard_replication_factor TO 1;
 
-CREATE TABLE user_info_data (user_id int, u_data user_data);
+CREATE TABLE user_info_data (user_id int, u_data user_data, user_index int);
 SELECT create_distributed_table('user_info_data', 'user_id');
 
 -- show that local id is 0, we'll use this information
@@ -46,10 +46,10 @@ SELECT get_local_node_id_stable();
 
 
 
-INSERT INTO user_info_data SELECT i, ('name' || i, i % 20 + 20)::user_data FROM generate_series(0,7)i;
+INSERT INTO user_info_data SELECT i, ('name' || i, i % 20 + 20)::user_data, i FROM generate_series(0,7)i;
 
 -- make sure that it is also true for  fast-path router queries with paramaters
-PREPARE fast_path_router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 RETURNING *;
+PREPARE fast_path_router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 RETURNING user_id, u_data;
 
 execute fast_path_router_with_param(0);
 execute fast_path_router_with_param(1);
@@ -62,7 +62,7 @@ execute fast_path_router_with_param(7);
 
 
 -- make sure that it is also true for  fast-path router queries with paramaters
-PREPARE fast_path_router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('test', get_local_node_id_stable())::user_data AND user_id  = $1 RETURNING *;
+PREPARE fast_path_router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('test', get_local_node_id_stable())::user_data AND user_id  = $1 RETURNING user_id, u_data;
 
 INSERT INTO user_info_data SELECT i, ('test', 0)::user_data FROM generate_series(0,7)i;
 
@@ -79,7 +79,7 @@ execute fast_path_router_with_param_and_func(7);
 
 INSERT INTO user_info_data SELECT 1, ('test' || i, 0)::user_data FROM generate_series(0,7)i;
 
-PREPARE fast_path_router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 1 RETURNING *;
+PREPARE fast_path_router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 1 RETURNING user_id, u_data;
 EXECUTE fast_path_router_with_param_on_non_dist_key_and_func(('test0', get_local_node_id_stable())::user_data);
 EXECUTE fast_path_router_with_param_on_non_dist_key_and_func(('test1', get_local_node_id_stable())::user_data);
 EXECUTE fast_path_router_with_param_on_non_dist_key_and_func(('test2', get_local_node_id_stable())::user_data);
@@ -93,7 +93,7 @@ EXECUTE fast_path_router_with_param_on_non_dist_key_and_func(('test7', get_local
 INSERT INTO user_info_data SELECT 1, ('test', i)::user_data FROM generate_series(0,7)i;
 
 PREPARE fast_path_router_with_param_on_non_dist_key(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 1 RETURNING
- *;
+ user_id, u_data;
 EXECUTE fast_path_router_with_param_on_non_dist_key(('test', 0)::user_data);
 EXECUTE fast_path_router_with_param_on_non_dist_key(('test', 1)::user_data);
 EXECUTE fast_path_router_with_param_on_non_dist_key(('test', 2)::user_data);
@@ -107,7 +107,7 @@ EXECUTE fast_path_router_with_param_on_non_dist_key(('test', 7)::user_data);
 INSERT INTO user_info_data SELECT i, ('test', i)::user_data FROM generate_series(0,7)i;
 
 PREPARE fast_path_router_with_two_params(user_data, int) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = $2 RETURNING
- *;
+ user_id, u_data;
 
 EXECUTE fast_path_router_with_two_params(('test', 0)::user_data, 0);
 EXECUTE fast_path_router_with_two_params(('test', 1)::user_data, 1);
@@ -121,7 +121,7 @@ EXECUTE fast_path_router_with_two_params(('test', 7)::user_data, 7);
 
 INSERT INTO user_info_data VALUES(1, ('test', 1)::user_data);
 
-PREPARE fast_path_router_with_only_function AS DELETE FROM user_info_data WHERE get_local_node_id_stable() = 0 AND user_id = 1 RETURNING *;
+PREPARE fast_path_router_with_only_function AS DELETE FROM user_info_data WHERE get_local_node_id_stable() = 0 AND user_id = 1 RETURNING user_id, u_data;
 EXECUTE fast_path_router_with_only_function;
 INSERT INTO user_info_data VALUES(1, ('test', 1)::user_data);
 EXECUTE fast_path_router_with_only_function;
@@ -137,8 +137,6 @@ INSERT INTO user_info_data VALUES(1, ('test', 1)::user_data);
 EXECUTE fast_path_router_with_only_function;
 INSERT INTO user_info_data VALUES(1, ('test', 1)::user_data);
 EXECUTE fast_path_router_with_only_function;
-
-ALTER TABLE user_info_data ADD COLUMN user_index INT;
 
 PREPARE insert_with_function_and_param(user_data) AS INSERT INTO user_info_data VALUES (1, $1, get_local_node_id_stable()) RETURNING user_id;
 EXECUTE insert_with_function_and_param(('test', 1)::user_data);
@@ -149,13 +147,12 @@ EXECUTE insert_with_function_and_param(('test', 1)::user_data);
 EXECUTE insert_with_function_and_param(('test', 1)::user_data);
 EXECUTE insert_with_function_and_param(('test', 1)::user_data);
 
-ALTER TABLE user_info_data DROP COLUMN user_index;
 TRUNCATE user_info_data;
 
 INSERT INTO user_info_data SELECT i, ('test', i)::user_data FROM generate_series(0,7)i;
 
 -- make sure that it is also true for non fast-path router queries with paramaters
-PREPARE router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 AND user_id = $1 RETURNING *;
+PREPARE router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 AND user_id = $1 RETURNING user_id, u_data;
 
 execute router_with_param(0);
 execute router_with_param(1);
@@ -168,7 +165,7 @@ execute router_with_param(7);
 
 
 -- make sure that it is also true for non fast-path router queries with paramaters
-PREPARE router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('test', get_local_node_id_stable())::user_data AND user_id  = $1 AND user_id  = $1 RETURNING *;
+PREPARE router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('test', get_local_node_id_stable())::user_data AND user_id  = $1 AND user_id  = $1 RETURNING user_id, u_data;
 
 INSERT INTO user_info_data SELECT i, ('test', 0)::user_data FROM generate_series(0,7)i;
 
@@ -184,7 +181,7 @@ execute router_with_param_and_func(7);
 
 INSERT INTO user_info_data SELECT 1, ('test' || i, 0)::user_data FROM generate_series(0,7)i;
 
-PREPARE router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id = 1 AND user_id  = 1 RETURNING *;
+PREPARE router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id = 1 AND user_id  = 1 RETURNING user_id, u_data;
 EXECUTE router_with_param_on_non_dist_key_and_func(('test0', get_local_node_id_stable())::user_data);
 EXECUTE router_with_param_on_non_dist_key_and_func(('test1', get_local_node_id_stable())::user_data);
 EXECUTE router_with_param_on_non_dist_key_and_func(('test2', get_local_node_id_stable())::user_data);
@@ -197,7 +194,7 @@ EXECUTE router_with_param_on_non_dist_key_and_func(('test7', get_local_node_id_s
 
 INSERT INTO user_info_data SELECT 1, ('test', i)::user_data FROM generate_series(0,7)i;
 
-PREPARE router_with_param_on_non_dist_key(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 1 AND user_id  = 1 RETURNING *;
+PREPARE router_with_param_on_non_dist_key(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 1 AND user_id  = 1 RETURNING user_id, u_data;
 EXECUTE router_with_param_on_non_dist_key(('test', 0)::user_data);
 EXECUTE router_with_param_on_non_dist_key(('test', 1)::user_data);
 EXECUTE router_with_param_on_non_dist_key(('test', 2)::user_data);
@@ -212,7 +209,7 @@ EXECUTE router_with_param_on_non_dist_key(('test', 7)::user_data);
 INSERT INTO user_info_data SELECT i, ('test', i)::user_data FROM generate_series(0,7)i;
 
 PREPARE router_with_two_params(user_data, int) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = $2 AND user_id  = $2 RETURNING
- *;
+ user_id, u_data;
 
 EXECUTE router_with_two_params(('test', 0)::user_data, 0);
 EXECUTE router_with_two_params(('test', 1)::user_data, 1);
@@ -225,7 +222,7 @@ EXECUTE router_with_two_params(('test', 7)::user_data, 7);
 
 INSERT INTO user_info_data VALUES(1, ('test', 1)::user_data);
 
-PREPARE router_with_only_function AS DELETE FROM user_info_data WHERE get_local_node_id_stable() = 0 AND user_id = 1 AND user_id = 1 RETURNING *;
+PREPARE router_with_only_function AS DELETE FROM user_info_data WHERE get_local_node_id_stable() = 0 AND user_id = 1 AND user_id = 1 RETURNING user_id, u_data;
 EXECUTE router_with_only_function;
 INSERT INTO user_info_data VALUES(1, ('test', 1)::user_data);
 EXECUTE router_with_only_function;
@@ -263,7 +260,7 @@ INSERT INTO user_info_data (user_id, u_data) VALUES
 
 
 -- make sure that it is also true for  fast-path router queries with paramaters
-PREPARE fast_path_router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 RETURNING *;
+PREPARE fast_path_router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 RETURNING user_id, u_data;
 
 execute fast_path_router_with_param(3);
 execute fast_path_router_with_param(4);
@@ -281,7 +278,7 @@ INSERT INTO user_info_data (user_id, u_data) VALUES
 (14, '(''test'', 2)'), (16, '(''test'', 2)');
 
 -- make sure that it is also true for  fast-path router queries with paramaters
-PREPARE fast_path_router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('''test''', get_constant_stable())::user_data AND user_id  = $1 RETURNING *;
+PREPARE fast_path_router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('''test''', get_constant_stable())::user_data AND user_id  = $1 RETURNING user_id, u_data;
 
 execute fast_path_router_with_param_and_func(3);
 execute fast_path_router_with_param_and_func(4);
@@ -292,7 +289,7 @@ execute fast_path_router_with_param_and_func(12);
 execute fast_path_router_with_param_and_func(14);
 execute fast_path_router_with_param_and_func(16);
 
-PREPARE fast_path_router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 RETURNING *;
+PREPARE fast_path_router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 RETURNING user_id, u_data;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, '(''test'', 2)'::user_data);
 EXECUTE fast_path_router_with_param_on_non_dist_key_and_func(('''test''', get_constant_stable())::user_data);
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, '(''test'', 2)');
@@ -311,7 +308,7 @@ EXECUTE fast_path_router_with_param_on_non_dist_key_and_func(('''test''', get_co
 
 
 
-PREPARE fast_path_router_with_param_on_non_dist_key(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 RETURNING *;
+PREPARE fast_path_router_with_param_on_non_dist_key(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 RETURNING user_id, u_data;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 1)::user_data);
 EXECUTE fast_path_router_with_param_on_non_dist_key(('test', 1)::user_data);
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 1)::user_data);
@@ -334,7 +331,7 @@ INSERT INTO user_info_data (user_id, u_data) VALUES
 (9, ('test', 2)), (11, ('test', 2)), (12, ('test', 2)),
 (14, ('test', 2)), (16, ('test', 2));
 
-PREPARE fast_path_router_with_two_params(user_data, int) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = $2 RETURNING *;
+PREPARE fast_path_router_with_two_params(user_data, int) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = $2 RETURNING user_id, u_data;
 
 EXECUTE fast_path_router_with_two_params(('test', 2)::user_data, 3);
 EXECUTE fast_path_router_with_two_params(('test', 2)::user_data, 4);
@@ -346,7 +343,7 @@ EXECUTE fast_path_router_with_two_params(('test', 2)::user_data, 14);
 EXECUTE fast_path_router_with_two_params(('test', 2)::user_data, 16);
 
 
-PREPARE fast_path_router_with_only_function AS DELETE FROM user_info_data WHERE get_constant_stable() = 2AND user_id = 3 RETURNING *;
+PREPARE fast_path_router_with_only_function AS DELETE FROM user_info_data WHERE get_constant_stable() = 2AND user_id = 3 RETURNING user_id, u_data;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 2)::user_data);
 EXECUTE fast_path_router_with_only_function;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 2)::user_data);
@@ -362,14 +359,6 @@ EXECUTE fast_path_router_with_only_function;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 2)::user_data);
 EXECUTE fast_path_router_with_only_function;
 
-
--------
-\c - - - :master_port
-SET search_path TO master_evaluation_combinations_modify;
-ALTER TABLE user_info_data ADD COLUMN user_index INT;
-
-\c - - - :worker_2_port
-SET search_path TO master_evaluation_combinations_modify;
 
 PREPARE insert_with_function_and_param(user_data) AS INSERT INTO user_info_data VALUES (3, $1, get_local_node_id_stable()) RETURNING user_id;
 EXECUTE insert_with_function_and_param(('test', 1)::user_data);
@@ -380,15 +369,7 @@ EXECUTE insert_with_function_and_param(('test', 1)::user_data);
 EXECUTE insert_with_function_and_param(('test', 1)::user_data);
 EXECUTE insert_with_function_and_param(('test', 1)::user_data);
 
-\c - - - :master_port
-SET search_path TO master_evaluation_combinations_modify;
-ALTER TABLE user_info_data DROP COLUMN user_index;
 TRUNCATE user_info_data;
-
-
-\c - - - :worker_2_port
-SET citus.log_local_commands TO ON;
-SET search_path TO master_evaluation_combinations_modify;
 
 -- all local values
 INSERT INTO user_info_data (user_id, u_data) VALUES
@@ -397,7 +378,7 @@ INSERT INTO user_info_data (user_id, u_data) VALUES
 (14, '(''test14'', 14)'), (16, '(''test16'', 16)');
 
 -- make sure that it is also true for  fast-path router queries with paramaters
-PREPARE router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 AND user_id = $1 RETURNING *;
+PREPARE router_with_param(int) AS DELETE FROM user_info_data WHERE user_id  = $1 AND user_id = $1 RETURNING user_id, u_data;
 
 execute router_with_param(3);
 execute router_with_param(4);
@@ -415,7 +396,7 @@ INSERT INTO user_info_data (user_id, u_data) VALUES
 (14, '(''test'', 2)'), (16, '(''test'', 2)');
 
 -- make sure that it is also true for  fast-path router queries with paramaters
-PREPARE router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('''test''', get_constant_stable())::user_data AND user_id  = $1 AND user_id  = $1 RETURNING *;
+PREPARE router_with_param_and_func(int) AS DELETE FROM user_info_data WHERE u_data = ('''test''', get_constant_stable())::user_data AND user_id  = $1 AND user_id  = $1 RETURNING user_id, u_data;
 
 execute router_with_param_and_func(3);
 execute router_with_param_and_func(4);
@@ -426,7 +407,7 @@ execute router_with_param_and_func(12);
 execute router_with_param_and_func(14);
 execute router_with_param_and_func(16);
 
-PREPARE router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 AND user_id  = 3 RETURNING *;
+PREPARE router_with_param_on_non_dist_key_and_func(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 AND user_id  = 3 RETURNING user_id, u_data;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, '(''test'', 2)'::user_data);
 EXECUTE router_with_param_on_non_dist_key_and_func(('''test''', get_constant_stable())::user_data);
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, '(''test'', 2)');
@@ -445,7 +426,7 @@ EXECUTE router_with_param_on_non_dist_key_and_func(('''test''', get_constant_sta
 
 
 
-PREPARE router_with_param_on_non_dist_key(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 AND user_id  = 3 RETURNING *;
+PREPARE router_with_param_on_non_dist_key(user_data) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = 3 AND user_id  = 3 RETURNING user_id, u_data;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 1)::user_data);
 EXECUTE router_with_param_on_non_dist_key(('test', 1)::user_data);
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 1)::user_data);
@@ -468,7 +449,7 @@ INSERT INTO user_info_data (user_id, u_data) VALUES
 (9, ('test', 2)), (11, ('test', 2)), (12, ('test', 2)),
 (14, ('test', 2)), (16, ('test', 2));
 
-PREPARE router_with_two_params(user_data, int) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = $2 AND user_id  = $2 RETURNING *;
+PREPARE router_with_two_params(user_data, int) AS DELETE FROM user_info_data WHERE u_data = $1 AND user_id  = $2 AND user_id  = $2 RETURNING user_id, u_data;
 
 EXECUTE router_with_two_params(('test', 2)::user_data, 3);
 EXECUTE router_with_two_params(('test', 2)::user_data, 4);
@@ -480,7 +461,7 @@ EXECUTE router_with_two_params(('test', 2)::user_data, 14);
 EXECUTE router_with_two_params(('test', 2)::user_data, 16);
 
 
-PREPARE router_with_only_function AS DELETE FROM user_info_data WHERE get_constant_stable() = 2AND user_id = 3 RETURNING *;
+PREPARE router_with_only_function AS DELETE FROM user_info_data WHERE get_constant_stable() = 2AND user_id = 3 RETURNING user_id, u_data;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 2)::user_data);
 EXECUTE router_with_only_function;
 INSERT INTO user_info_data (user_id, u_data) VALUES  (3, ('test', 2)::user_data);
